Add strength.
Strength is a Runner stat that determines Carrying Capacity when Ingredient Running.

10 Strength equates to +1 capacity.

Strength is limited in a range from 1-50.

Any remainder strength not a multiple of 10 will be rolled as a percentile to gain +1 strength.

E.g - 7 strength = 70% chance to get a bonus on top of the other capacity generated from other strength, e.g 40 if user's strength was 47