Add Focus as a Player Stat.
Focus is a stat that determines how much a player will "focus" on one ingredient. Every point in Focus is a 1% chance during Ingredient Selection for a Chef to choose the same ingredient that they chose last turn.

If no ingredient was selected last turn (e.g it is the first turn) or Focus does not proc, the next ingredient will be selected at random, which can be the same ingredient as the Last Ingredient, just randomly selected.

Focus is naturally capped at 50, for 50% focus chance.